fix: jpa, schema 정합성 (#367)
* refactor: activity id Integer로 변경

* refactor: timestamp 명시

* refactor: LocalDateTimeAttributeConverter 개선

* refactor: id bigint로 변경

* Revert "refactor: activity id Integer로 변경"

This reverts commit bea44bb4c28c7e26b7494fcface5f8e95269f9cb.

* refactor: LocalDateTime 컬럼에 대해 columnDefinition 추가

* refactor: id 수정 flyway 제거

* refactor: Activity id Long -> Integer 수정

* refactor: Article id Long -> Integer 수정

* refactor: User id Long -> Integer 수정

* refactor: Article content, hit 수정

* refactor: Board id Long -> Integer 수정

* refactor: Board articleCount Long -> Integer 수정

* refactor: Board parentId, seq Long -> Integer 수정

* refactor: Comment id, content 수정

* refactor: Dining id, date 수정

* refactor: Dining soldOut, isChanged 수정

* refactor: EventArticle id, content 수정

* refactor: Shop, Menu Entity validate 맞게 수정

* refactor: Land validate 맞게 수정

* refactor: Lecture id Long -> Integer 수정

* refactor: Member, Track, Board Notification validate 조건에 맞게 수정

* refactor: Notification 수정

* refactor: OwnerAttachment 수정

* refactor: Owner boolean 처리

* refactor: Semester id Long -> Integer 수정

* refactor: ShopCategory id Long -> Integer 수정

* refactor: ShopCategoryMap id Long -> Integer 수정

* refactor: ShopImage id Long -> Integer 수정

* refactor: MenuOption id Long -> Integer 수정

* refactor: MenuImage, ShopOpen id Long -> Integer 수정

* refactor: Boolean 값 primitive로 수정

* refactor: Shop 필드값 수정

* refactor: Student identity 수정

* refactor: TechStack imageUrl 수정

* refactor: User 정보 수정

* refactor: Repository 정보 수정

* test: 테스트 수정

+++ src/main/java/in/koreatech/koin/domain/community/repository/BoardRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.repository.Repository;
 import in.koreatech.koin.domain.community.exception.ArticleNotFoundException;
 import in.koreatech.koin.domain.community.model.Board;
 
-public interface BoardRepository extends Repository<Board, Long> {
-    Optional<Board> findById(Long id);
+public interface BoardRepository extends Repository<Board, Integer> {
+    Optional<Board> findById(Integer id);
 
     Board save(Board board);
 
-    default Board getById(Long boardId) {
+    default Board getById(Integer boardId) {
         return findById(boardId).orElseThrow(
             () -> ArticleNotFoundException.withDetail("boardId: " + boardId));
     }
