+++ src/main/java/in/koreatech/koin/domain/shop/repository/MenuCategoryMapRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.Repository;
 
 import in.koreatech.koin.domain.shop.model.MenuCategoryMap;
 
-public interface MenuCategoryMapRepository extends Repository<MenuCategoryMap, Long> {
+public interface MenuCategoryMapRepository extends Repository<MenuCategoryMap, Integer> {
 
     MenuCategoryMap save(MenuCategoryMap menuCategoryMap);
 }
