+++ src/main/java/in/koreatech/koin/domain/shop/repository/ShopCategoryMapRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.Repository;
 
 import in.koreatech.koin.domain.shop.model.ShopCategoryMap;
 
-public interface ShopCategoryMapRepository extends Repository<ShopCategoryMap, Long> {
+public interface ShopCategoryMapRepository extends Repository<ShopCategoryMap, Integer> {
 
     ShopCategoryMap save(ShopCategoryMap shopCategoryMap);
 
-    List<ShopCategoryMap> findAllByShopId(Long shopId);
+    List<ShopCategoryMap> findAllByShopId(Integer shopId);
 }
