+++ src/main/java/in/koreatech/koin/domain/shop/repository/ShopCategoryRepository.java
@@ -8,15 +8,15 @@ import org.springframework.data.repository.Repository;
 import in.koreatech.koin.domain.shop.exception.ShopCategoryNotFoundException;
 import in.koreatech.koin.domain.shop.model.ShopCategory;
 
-public interface ShopCategoryRepository extends Repository<ShopCategory, Long> {
+public interface ShopCategoryRepository extends Repository<ShopCategory, Integer> {
 
-    Optional<ShopCategory> findById(Long shopCategoryId);
+    Optional<ShopCategory> findById(Integer shopCategoryId);
 
     ShopCategory save(ShopCategory shopCategory);
 
-    List<ShopCategory> findAllByIdIn(List<Long> ids);
+    List<ShopCategory> findAllByIdIn(List<Integer> ids);
 
-    default ShopCategory getById(Long shopCategoryId) {
+    default ShopCategory getById(Integer shopCategoryId) {
         return findById(shopCategoryId)
             .orElseThrow(() -> ShopCategoryNotFoundException.withDetail("shopCategoryId: " + shopCategoryId));
     }
